+++ src/longToWide.cpp
@@ -16,6 +16,7 @@
 // along with icd. If not, see <http://www.gnu.org/licenses/>.
 
 #include "icd_types.h"                  // for VecStr, VecVecStr, CV
+#include "local.h"
 #include <string.h>                     // for strcmp
 #include <algorithm>                    // for find, fill
 #include <string>                       // for string, operator==
@@ -26,137 +27,90 @@
 #endif
 
 extern "C" {
-#include "cutil.h"                            // for getRListOrDfElement
+#include "cutil.h"
 }
 
 CV raggedToWide(const VecVecStr& ragged, int max_per_pt,
-                const VecStr &visitIds) {
+                const VecStr &ids) {
 #ifdef ICD_DEBUG_TRACE
-  Rcpp::Rcout << "visitIds = ";
-  // printIt(visitIds); // broken, not sure why.
+  Rcpp::Rcout << "ids = ";
+  // printIt(ids); // broken, not sure why.
 #endif
   VecStr::size_type distinct_visits = ragged.size();
-  CV out(distinct_visits * max_per_pt, NA_STRING); // optionally default empty strings? NA? User can do this for now.
+  // optionally default empty strings? NA? User can do this for now.
+  CV out(distinct_visits * max_per_pt, NA_STRING);
 #ifdef ICD_DEBUG
-  if (distinct_visits == 0) {
-    Rcpp::Rcout << "no visits. returning blank data\n";
-    return CV::create();
-  }
-  if (distinct_visits != visitIds.size()) {
-    Rcpp::Rcout << "visit and ragged sizes differ. visits = " << visitIds.size() <<
-      ", ragged size = " << distinct_visits << ": returning blank data\n";
-    return CV::create();
-  }
+  assert(distinct_visits > 0)
+    assert(distinct_visits == ids.size())
 #endif
-  for (VecVecStr::size_type row_it = 0; row_it < distinct_visits; ++row_it) {
-    const VecStr& this_row = ragged[row_it];
-    VecStr::size_type this_row_len = this_row.size();
-    for (VecStr::size_type col_it = 0; col_it < this_row_len; ++col_it) {
-      // write in row major format, but this means transpose needed later
-      VecVecStr::size_type out_idx = row_it + (distinct_visits * col_it);
-      out[out_idx] = this_row[col_it];
-    } // end inner for
-    Rcpp::checkUserInterrupt();
-  } // end outer for
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "writing dimensions\n";
-#endif
-  // set dimensions in reverse (row major for parallel step)
-  out.attr("dim") = Rcpp::Dimension(distinct_visits, max_per_pt);
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "writing labels\n";
-#endif
+    for (VecVecStr::size_type row_it = 0; row_it < distinct_visits; ++row_it) {
+      const VecStr& this_row = ragged[row_it];
+      VecStr::size_type this_row_len = this_row.size();
+      for (VecStr::size_type col_it = 0; col_it < this_row_len; ++col_it) {
+        // write in row major format, but this means transpose needed later
+        VecVecStr::size_type out_idx = row_it + (distinct_visits * col_it);
+        out[out_idx] = this_row[col_it];
+      } // end inner for
+      Rcpp::checkUserInterrupt();
+    } // end outer for
+    // set dimensions in reverse (row major for parallel step)
+    out.attr("dim") = Rcpp::Dimension(distinct_visits, max_per_pt);
   CV nonames;
-  rownames(out) = Rcpp::wrap(visitIds);
+  rownames(out) = Rcpp::wrap(ids);
   return out;
 }
 
-int longToRagged(const SEXP& icd9df, VecVecStr& ragged, VecStr& visitIds,
-                 const std::string visitId, const std::string icd9Field =
+int longToRagged(const SEXP& x, VecVecStr& ragged, VecStr& ids,
+                 const std::string id_name, const std::string code_name =
                    "code", bool aggregate = true) {
-#ifdef ICD_VALGRIND
-  CALLGRIND_START_INSTRUMENTATION;
-#endif
-  SEXP icds = PROTECT(getRListOrDfElement(icd9df, icd9Field.c_str()));
-  SEXP vsexp = PROTECT(getRListOrDfElement(icd9df, visitId.c_str()));
+  SEXP icds = PROTECT(getRListOrDfElement(x, code_name.c_str()));
+  SEXP vsexp = PROTECT(getRListOrDfElement(x, id_name.c_str()));
   const int approx_cmb_per_visit = 15; // just an estimate. Prob best to overestimate.
   int vlen = Rf_length(icds);
-  visitIds.reserve(vlen / approx_cmb_per_visit);
+  ids.reserve(vlen / approx_cmb_per_visit);
   ragged.reserve(vlen / approx_cmb_per_visit);
   int max_per_pt = 1;
-
-  const char* lastVisitId = "";
+  const char* lastid_name = "";
   for (int i = 0; i < vlen; ++i) {
     // always STRING? may get numeric, integer, factor? Can always handle this on R side
     const char* icd = CHAR(STRING_ELT(icds, i));
     const char* vi = CHAR(STRING_ELT(vsexp, i));
-
-    if (strcmp(lastVisitId, vi) != 0
+    if (strcmp(lastid_name, vi) != 0
           && (!aggregate
-                || std::find(visitIds.rbegin(), visitIds.rend(), vi)
-                == visitIds.rend())) {
+                || std::find(ids.rbegin(), ids.rend(), vi)
+                == ids.rend())) {
                 VecStr vcodes;
       vcodes.reserve(approx_cmb_per_visit); // estimate of number of codes per patient.
       vcodes.push_back(icd); // new vector of ICD codes with this first item
       ragged.push_back(vcodes); // and add that vector to the intermediate structure
-      visitIds.push_back(vi);
+      ids.push_back(vi);
     } else {
-#ifdef ICD_DEBUG
-      if (ragged.size() == 0) {
-        Rcpp::Rcout << "ragged size is ZERO! bailing ou!\n";
-        break;
-      }
-#endif
+      assert(ragged.size() != 0);
       ragged[ragged.size() - 1].push_back(icd); // augment vec for current visit and N/V/E type // EXPENSIVE.
       int len = ragged[ragged.size() - 1].size(); // get new count of cmb for one patient
       if (len > max_per_pt)
         max_per_pt = len;
     }
-#ifdef ICD_DEBUG_TRACE
-    Rcpp::Rcout << "ragged size is " << ragged.size() << "\n";
-#endif
-    lastVisitId = vi;
+    lastid_name = vi;
     Rcpp::checkUserInterrupt();
   } // end loop through all visit-code input data
-
-#ifdef ICD_VALGRIND
-  CALLGRIND_STOP_INSTRUMENTATION;
-  //        CALLGRIND_DUMP_STATS;
-#endif
   UNPROTECT(2);
   return max_per_pt;
 }
 
 // [[Rcpp::export(long_to_wide_cpp)]]
-CV longToWideCpp(const SEXP& icd9df,
-                 const std::string visitId, const std::string icd9Field,
-                 bool aggregate = true) {
-
-  // a few options here. character matrix would make sense, but this can't be a factor. data frame harder to process on C++ side.
-  // Matrix easy to convert to data frame in R, if needed. Rows in output don't correspond to input, so no strong reason to preserve factors in any way.
-  // therefore, we will only accept strings as input at the C++ level. R wrapper function can pre-convert these.
-
+CV longToWideCpp(const SEXP& x, const std::string id_name,
+                 const std::string code_name, bool aggregate = true) {
   VecVecStr ragged; // reserve size later
   int max_per_pt = 0;
-
-  const SEXP vsexp = PROTECT(getRListOrDfElement(icd9df, visitId.c_str()));
+  const SEXP vsexp = PROTECT(getRListOrDfElement(x, id_name.c_str()));
   if (TYPEOF(vsexp) != STRSXP)
     Rcpp::stop(
       "The contents of the visit ID column should be converted to character.");
   UNPROTECT(1);
-
-  VecStr visitIds; // may be vector of integers or strings
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "doing long to ragged\n";
-#endif
-  max_per_pt = longToRagged(icd9df, ragged, visitIds, visitId, icd9Field,
+  VecStr ids; // may be vector of integers or strings
+  max_per_pt = longToRagged(x, ragged, ids, id_name, code_name,
                             aggregate);
-#ifdef ICD_DEBUG_TRACE
-  Rcpp::Rcout << "ragged size returned is " << ragged.size() << "\n";
-#endif
-
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "returning ragged to wide\n";
-#endif
-  return raggedToWide(ragged, max_per_pt, visitIds);
+  DEBUG_VEC(ragged);
+  return raggedToWide(ragged, max_per_pt, ids);
 }